backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
psycopg2-binary
pgvector
xgboost
fpdf2
sqlalchemy>=2.0
python-dotenv
groq
apscheduler
httpx
python-docx
pdfplumber
pypdf
scikit-learn
plotly
pandas
numpy
langgraph>=0.2.0
langchain-core>=0.3.0
pydantic-settings>=2.0.0
python-multipart
celery[redis]>=5.3.0
redis>=5.0.0
qdrant-client>=1.9.0
pydantic>=2.0.0
